refactor, use $PropertyType flowtype util
