+++ src/InputOrder/type.flow.js
@@ -1,0 +1,6 @@
+// @flow
+export type Value = number | string;
+export type ItemProps = {
+  value: Value,
+  children: React.Node,
+};
